+++ rugosa/__init__.py
@@ -1,4 +1,4 @@
 from .emulation import Emulator
 from .disassembly import *
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"