+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.2.0] - 2022-06-01
 
 ### Changed
 - IDA: Greatly improved emulation performance by "teleporting" the `Emulator` instance into the IDA interpreter.
@@ -29,5 +29,6 @@ All notable changes to this project will be documented in this file.
 - Migrated the majority of Kordesii functionality to work with Dragodis.
 
 
-[Unreleased]: https://github.com/Defense-Cyber-Crime-Center/rugosa/compare/0.1.1...HEAD
-[0.1.1]: https://github.com/Defense-Cyber-Crime-Center/rugosa/compare/0.1.0...0.1.1
+[Unreleased]: https://github.com/dod-cyber-crime-center/rugosa/compare/0.1.1...HEAD
+[0.2.0]: https://github.com/dod-cyber-crime-center/rugosa/compare/0.1.1...0.2.0
+[0.1.1]: https://github.com/dod-cyber-crime-center/rugosa/compare/0.1.0...0.1.1